Source information is working for .NET 4.5.1
