+++ src/dotnet-test-nunit/project.json
@@ -23,6 +23,7 @@
 
   "dependencies": {
     "Microsoft.Extensions.Testing.Abstractions": "1.0.0-preview1-002702",
+    "Microsoft.Extensions.DependencyModel": "1.0.0-rc2-final",
     "NUnit.Portable.Agent": "3.3.0-CI-2746-master",
     "NETStandard.Library": "1.5.0-rc2-24027"
   },
@@ -38,7 +39,8 @@
           "type": "platform",
           "version": "1.0.0-rc2-3002702"
         },
-        "System.ObjectModel": "4.0.12-rc2-24027"
+        "System.ObjectModel": "4.0.12-rc2-24027",
+        "System.Runtime.Loader": "4.0.0-rc2-24027"
       }
     },
     "net451": {
